backend/requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cors==1.0.1
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
future==1.0.0
gevent==25.5.1
google-auth==2.40.1
google-auth-oauthlib==1.2.2
greenlet==3.2.2
gspread==6.2.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauth2client==4.1.3
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tldextract==5.3.0
urllib3==2.4.0
Werkzeug==3.1.3
xmltodict==0.14.2
zope.event==5.0
zope.interface==7.2
